# Optimized Lagrange Interpolation