--- a/sympy/polys/specialpolys.py
+++ b/sympy/polys/specialpolys.py
@@ -116,7 +116,7 @@ def symmetric_poly(n, *gens, **args):
     elif not n:
         poly = S.One
     else:
-        poly = Add(*[ Mul(*s) for s in subsets(gens, int(n)) ])
+        poly = Add(*[Mul(*s) for s in subsets(gens, int(n))])
 
     if not args.get('polys', False):
         return poly
@@ -161,24 +161,24 @@ def interpolating_poly(n, x, X='x', Y='y'):
         Y = symbols("%s:%s" % (Y, n))
 
     coeffs = []
-    numert = Mul(*[ (x - u) for u in X ])
+    numert = Mul(*[(x - u) for u in X])
 
-    for i in range(0, n):
+    for i in range(n):
         numer = numert/(x - X[i])
-        denom = Mul(*[ (X[i] - X[j]) for j in range(0, n) if i != j ])
+        denom = Mul(*[(X[i] - X[j]) for j in range(n) if i != j])
         coeffs.append(numer/denom)
 
-    return Add(*[ coeff*y for coeff, y in zip(coeffs, Y) ])
+    return Add(*[coeff*y for coeff, y in zip(coeffs, Y)])
 
 
 def fateman_poly_F_1(n):
     """Fateman's GCD benchmark: trivial GCD """
-    Y = [ Symbol('y_' + str(i)) for i in range(0, n + 1) ]
+    Y = [Symbol('y_' + str(i)) for i in range(n + 1)]
 
     y_0, y_1 = Y[0], Y[1]
 
-    u = y_0 + Add(*[ y for y in Y[1:] ])
-    v = y_0**2 + Add(*[ y**2 for y in Y[1:] ])
+    u = y_0 + Add(*[y for y in Y[1:]])
+    v = y_0**2 + Add(*[y**2 for y in Y[1:]])
 
     F = ((u + 1)*(u + 2)).as_poly(*Y)
     G = ((v + 1)*(-3*y_1*y_0**2 + y_1**2 - 1)).as_poly(*Y)
@@ -192,7 +192,7 @@ def dmp_fateman_poly_F_1(n, K):
     """Fateman's GCD benchmark: trivial GCD """
     u = [K(1), K(0)]
 
-    for i in range(0, n):
+    for i in range(n):
         u = [dmp_one(i, K), u]
 
     v = [K(1), K(0), K(0)]
@@ -220,11 +220,11 @@ def dmp_fateman_poly_F_1(n, K):
 
 def fateman_poly_F_2(n):
     """Fateman's GCD benchmark: linearly dense quartic inputs """
-    Y = [ Symbol('y_' + str(i)) for i in range(0, n + 1) ]
+    Y = [Symbol('y_' + str(i)) for i in range(n + 1)]
 
     y_0 = Y[0]
 
-    u = Add(*[ y for y in Y[1:] ])
+    u = Add(*[y for y in Y[1:]])
 
     H = Poly((y_0 + u + 1)**2, *Y)
 
@@ -238,7 +238,7 @@ def dmp_fateman_poly_F_2(n, K):
     """Fateman's GCD benchmark: linearly dense quartic inputs """
     u = [K(1), K(0)]
 
-    for i in range(0, n - 1):
+    for i in range(n - 1):
         u = [dmp_one(i, K), u]
 
     m = n - 1
@@ -257,11 +257,11 @@ def dmp_fateman_poly_F_2(n, K):
 
 def fateman_poly_F_3(n):
     """Fateman's GCD benchmark: sparse inputs (deg f ~ vars f) """
-    Y = [ Symbol('y_' + str(i)) for i in range(0, n + 1) ]
+    Y = [Symbol('y_' + str(i)) for i in range(n + 1)]
 
     y_0 = Y[0]
 
-    u = Add(*[ y**(n + 1) for y in Y[1:] ])
+    u = Add(*[y**(n + 1) for y in Y[1:]])
 
     H = Poly((y_0**(n + 1) + u + 1)**2, *Y)