--- a/sympy/polys/tests/test_polyfuncs.py
+++ b/sympy/polys/tests/test_polyfuncs.py
@@ -8,7 +8,7 @@ from sympy.polys.polyerrors import (
     MultivariatePolynomialError,
 )
 
-from sympy import symbols
+from sympy import symbols, S
 from sympy.utilities.pytest import raises
 
 from sympy.abc import a, b, c, d, e, x, y, z
@@ -80,6 +80,10 @@ def test_interpolate():
     assert interpolate([(1, 1), (2, 4), (3, 9)], x) == x**2
     assert interpolate([(1, 2), (2, 5), (3, 10)], x) == 1 + x**2
     assert interpolate({1: 2, 2: 5, 3: 10}, x) == 1 + x**2
+    assert interpolate({5: 2, 7: 5, 8: 10, 9: 13}, x) == \
+        -S(13)*x**3/24 + S(12)*x**2 - S(2003)*x/24 + 187
+    assert interpolate([(1, 3), (0, 6), (2, 5), (5, 7), (-2, 4)], x) == \
+        S(-61)*x**4/280 + S(247)*x**3/210 + S(139)*x**2/280 - S(1871)*x/420 + 6
 
 
 def test_rational_interpolate():