--- a/sympy/polys/tests/test_specialpolys.py
+++ b/sympy/polys/tests/test_specialpolys.py
@@ -70,7 +70,7 @@ def test_random_poly():
 
 
 def test_interpolating_poly():
-    x0, x1, x2, y0, y1, y2 = symbols('x:3, y:3')
+    x0, x1, x2, x3, y0, y1, y2, y3 = symbols('x:4, y:4')
 
     assert interpolating_poly(0, x) == 0
     assert interpolating_poly(1, x) == y0
@@ -82,6 +82,12 @@ def test_interpolating_poly():
         y0*(x - x1)*(x - x2)/((x0 - x1)*(x0 - x2)) + \
         y1*(x - x0)*(x - x2)/((x1 - x0)*(x1 - x2)) + \
         y2*(x - x0)*(x - x1)/((x2 - x0)*(x2 - x1))
+
+    assert interpolating_poly(4, x) == \
+        y0*(x - x1)*(x - x2)*(x - x3)/((x0 - x1)*(x0 - x2)*(x0 - x3)) + \
+        y1*(x - x0)*(x - x2)*(x - x3)/((x1 - x0)*(x1 - x2)*(x1 - x3)) + \
+        y2*(x - x0)*(x - x1)*(x - x3)/((x2 - x0)*(x2 - x1)*(x2 - x3)) + \
+        y3*(x - x0)*(x - x1)*(x - x2)/((x3 - x0)*(x3 - x1)*(x3 - x2))
 
 
 def test_fateman_poly_F_1():