@jksuom 2018-04-08
Thanks, this looks good.